refactor: 컨트롤러에 존재하던 도메인 로직을 새로운 클래스로 이동, 불필요한 예외 처리 삭제
