--- src/main/kotlin/racingcar/exceptions/CarExceptions.kt
@@ -1,3 +1,0 @@
-package racingcar.exceptions
-
-class CarExceptions
