+++ src/main/kotlin/racingcar/exceptions/WorldExceptions.kt
@@ -10,8 +10,4 @@ class WorldExceptions {
     fun validateAttemptCount(attemptCount: Int) {
         require(attemptCount > 0) { "1번 이상 시도해아 합니다." }
     }
-
-    fun validateProcessStep(car: Car, number: Int) {
-        require(number in 0..9) { "생성된 임의 숫자는 0에서 9사이어야 합니다." }
-    }
 }
