+++ src/main/kotlin/racingcar/model/Car.kt
@@ -1,9 +1,6 @@
 package racingcar.model
 
-import racingcar.exceptions.CarExceptions
-
 class Car(private var name: String) {
-    private val exception = CarExceptions()
     private var position: Int = 0
 
     fun forward() {
